+++ benchmark/tests/rendering/bench.tsx
@@ -37,7 +37,7 @@ import {
 } from 'xstate'
 import { useSelector as useXSelector } from '@xstate/react'
 import { machine, type Machine } from '@dunky.dev/state-machine'
-import { useSelector } from '@dunky.dev/react-state-machine'
+import { useSelector } from '@dunky.dev/state-machine-react'
 
 type Ctx = { highlighted: number }
 type Ev = { type: 'move'; to: number }