+++ benchmark/CHANGELOG.md
@@ -1,5 +1,13 @@
 # benchmark
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @dunky.dev/state-machine-react@0.2.0
+  - @dunky.dev/state-machine@0.2.0
+
 ## 0.1.0
 
 ### Minor Changes